Bump FRB v2 and fix compilation due to breaking changes
