+++ .github/workflows/release-linux-appimage-x64.yml
@@ -61,7 +61,7 @@ jobs:
       uses: baptiste0928/cargo-install@v3
       with:
         crate: flutter_rust_bridge_codegen
-        version: "2.0.0-dev.35"
+        version: "2.0.0-dev.36"
 
     - name: Install lddtree
       uses: baptiste0928/cargo-install@v3
