+++ .github/workflows/release-macos-x64.yml
@@ -62,7 +62,7 @@ jobs:
       uses: baptiste0928/cargo-install@v3
       with:
         crate: flutter_rust_bridge_codegen
-        version: "2.0.0-dev.35"
+        version: "2.0.0-dev.36"
 
     - name: Install flutter_distributor
       run: dart pub global activate flutter_distributor
