+++ .github/workflows/release-win-x64.yml
@@ -75,7 +75,7 @@ jobs:
       uses: baptiste0928/cargo-install@v3
       with:
         crate: flutter_rust_bridge_codegen
-        version: "2.0.0-dev.35"
+        version: "2.0.0-dev.36"
         args: --target x86_64-pc-windows-msvc
 
     # - name: Install flutter_distributor
