+++ lib/views/photo_details_screen/photo_details_screen_view_model.dart
@@ -68,7 +68,7 @@ abstract class PhotoDetailsScreenViewModelBase extends ScreenViewModelBase with 
         StatsManager.instance.addUploadedPhoto();
       } else {
         logDebug("Uploading: ${event.transferredBytes}/${event.totalBytes} bytes");
-        _uploadProgress = event.transferredBytes / (event.totalBytes ?? 0);
+        _uploadProgress = event.transferredBytes / (event.totalBytes ?? BigInt.zero);
       }
     }).onError((x) async {
       logError("Upload failed, file path: ${file!.path}", x);
