+++ pubspec.yaml
@@ -51,7 +51,7 @@ dependencies:
   rust_lib_momento_booth:
     path: rust_builder
   ffi: 2.1.2
-  flutter_rust_bridge: 2.0.0-dev.35
+  flutter_rust_bridge: 2.0.0-dev.36
 
   # Printing
   printing: 5.12.0
