+++ rust/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 crate-type = ["cdylib", "staticlib"]
 
 [dependencies]
-flutter_rust_bridge = { version = "=2.0.0-dev.35", features = ["chrono"] }
+flutter_rust_bridge = { version = "=2.0.0-dev.36", features = ["chrono"] }
 anyhow = "1.0.86"
 derive_more = "0.99.17"
 atom_table = "1.1.0"
